# gen/http/openapi.yaml
swagger: "2.0"
info:
    title: 流程编辑器 API
    version: 0.0.1
host: localhost:80
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
    /account:
        get:
            tags:
                - account
            summary: getAccount account
            description: Get account
            operationId: account#getAccount
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/AccountGetAccountResponseBody'
                        required:
                            - activated
                            - membership
            schemes:
                - http
    /flow:
        get:
            tags:
                - flow
            summary: getFlowList flow
            description: 列出当前用户拥有的 flow
            operationId: flow#getFlowList
            responses:
                "200":
                    description: OK response.
                    schema:
                        type: array
                        items:
                            $ref: '#/definitions/FlowListItemResponse'
            schemes:
                - http
        post:
            tags:
                - flow
            summary: createFlow flow
            description: 创建 flow
            operationId: flow#createFlow
            parameters:
                - name: CreateFlowRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/FlowCreateFlowRequestBody'
                    required:
                        - title
            responses:
                "201":
                    description: Created response.
                    schema:
                        $ref: '#/definitions/FlowCreateFlowResponseBody'
                        required:
                            - id
                            - title
                            - created_at
            schemes:
                - http
    /flow/{id}:
        get:
            tags:
                - flow
            summary: getFlow flow
            description: 根据 flow id 获取 flow 详情
            operationId: flow#getFlow
            parameters:
                - name: id
                  in: path
                  required: true
                  type: string
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/FlowGetFlowResponseBody'
                        required:
                            - id
                            - title
                            - created_at
            schemes:
                - http
        put:
            tags:
                - flow
            summary: updateFlow flow
            description: 更新 flow
            operationId: flow#updateFlow
            parameters:
                - name: id
                  in: path
                  required: true
                  type: string
                - name: UpdateFlowRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/FlowUpdateFlowRequestBody'
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/FlowUpdateFlowResponseBody'
                        required:
                            - id
                            - title
                            - created_at
            schemes:
                - http
        delete:
            tags:
                - flow
            summary: deleteFlow flow
            description: 删除 flow
            operationId: flow#deleteFlow
            parameters:
                - name: id
                  in: path
                  required: true
                  type: string
            responses:
                "204":
                    description: No Content response.
            schemes:
                - http
definitions:
    AccountGetAccountResponseBody:
        title: AccountGetAccountResponseBody
        type: object
        properties:
            activated:
                type: boolean
                example: false
            membership:
                type: integer
                example: 1154100410672721326
                format: int64
        example:
            activated: false
            membership: 1317109992280416092
        required:
            - activated
            - membership
    FlowCreateFlowRequestBody:
        title: FlowCreateFlowRequestBody
        type: object
        properties:
            edges:
                type: string
                description: flow 边
                example: Blanditiis consequatur voluptate.
            nodes:
                type: string
                description: flow 节点
                example: Itaque et sed culpa vitae nemo minus.
            title:
                type: string
                description: flow 标题
                example: Nobis est deserunt blanditiis rerum.
        example:
            edges: Nesciunt alias labore.
            nodes: At in quibusdam reprehenderit.
            title: Consequatur qui deleniti et est.
        required:
            - title
    FlowCreateFlowResponseBody:
        title: FlowCreateFlowResponseBody
        type: object
        properties:
            created_at:
                type: string
                description: flow 创建时间
                example: Deleniti dolorum vel.
            edges:
                type: string
                description: flow 边
                example: Consequatur velit libero.
            id:
                type: integer
                description: flow id
                example: 1817256496796312647
                format: int64
            nodes:
                type: string
                description: flow 节点
                example: Et sequi qui molestiae aut omnis distinctio.
            title:
                type: string
                description: flow 标题
                example: Odio commodi consequatur et voluptatibus autem.
        example:
            created_at: Et voluptas quod possimus earum repudiandae.
            edges: Impedit quo nihil ut repellendus molestiae consequuntur.
            id: 3459883958734984567
            nodes: Ipsam laboriosam laudantium iusto pariatur.
            title: Illum possimus sit.
        required:
            - id
            - title
            - created_at
    FlowGetFlowResponseBody:
        title: FlowGetFlowResponseBody
        type: object
        properties:
            created_at:
                type: string
                description: flow 创建时间
                example: Accusamus saepe nemo.
            edges:
                type: string
                description: flow 边
                example: Nobis vitae eos.
            id:
                type: integer
                description: flow id
                example: 3501558131119642822
                format: int64
            nodes:
                type: string
                description: flow 节点
                example: Odit vero rerum qui sint id sint.
            title:
                type: string
                description: flow 标题
                example: Eos aut consequuntur.
        example:
            created_at: Velit voluptate temporibus vitae.
            edges: Similique ducimus.
            id: 8075908555921940998
            nodes: Aut doloribus alias inventore totam repudiandae.
            title: Praesentium fugiat maiores corporis consequatur ratione ab.
        required:
            - id
            - title
            - created_at
    FlowListItemResponse:
        title: FlowListItemResponse
        type: object
        properties:
            created_at:
                type: string
                description: flow 创建时间
                example: Ullam nisi sint voluptates cupiditate dignissimos.
            id:
                type: integer
                description: flow id
                example: 3045043540898249472
                format: int64
            title:
                type: string
                description: flow 标题
                example: Sunt ut minus dolores consequatur eius.
        example:
            created_at: Rerum rerum.
            id: 8028815304413024753
            title: Rerum similique est tempore fugit voluptatem.
        required:
            - id
            - title
            - created_at
    FlowUpdateFlowRequestBody:
        title: FlowUpdateFlowRequestBody
        type: object
        properties:
            edges:
                type: string
                description: flow 边
                example: Animi at optio ex quae.
            nodes:
                type: string
                description: flow 节点
                example: Distinctio dolorem reprehenderit.
            title:
                type: string
                description: flow 标题
                example: Veniam tempore.
        example:
            edges: Voluptas iure assumenda quos quae.
            nodes: Ut quis et repellat facere et.
            title: Modi dignissimos non perferendis possimus sint.
    FlowUpdateFlowResponseBody:
        title: FlowUpdateFlowResponseBody
        type: object
        properties:
            created_at:
                type: string
                description: flow 创建时间
                example: Est optio aperiam.
            edges:
                type: string
                description: flow 边
                example: Repellat corrupti reprehenderit.
            id:
                type: integer
                description: flow id
                example: 6882596758243384156
                format: int64
            nodes:
                type: string
                description: flow 节点
                example: Provident veniam porro ducimus voluptatem consequatur asperiores.
            title:
                type: string
                description: flow 标题
                example: Quis delectus.
        example:
            created_at: Facilis deserunt.
            edges: Magnam repudiandae error numquam.
            id: 7244976131529825890
            nodes: Iste ab.
            title: In itaque et.
        required:
            - id
            - title
            - created_at